Remove RationalQuasiIsolatedForms
The function as written picks a coset representative w of the
centraliser and then checks if s^w is in the group generated by s. This
isn't enough. One needs to check if the elements of the group generated
by s is in the W-orbit of s. This seems more computationally
challenging.